week5-agent-solution/requirements.txt
langchain==0.2.6
llama-index==0.10.14
openai==1.3.7
streamlit==1.31.1
python-dotenv==1.0.1
pypdf==4.2.0
docx2txt==0.8
